# docs/api.spec.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "JASPELKU FOR USER",
        "version": "1.0",
        "description": "Api specification for jaspelku project"
    },
    "servers": [
        {
            "url": "http:localhost:2008",
            "description": "development server"
        }
    ],
    "paths": {
...
        "/user/register" : {
            "post" : {
                "tags": ["user/auth"],
                "requestBody" : {
                    "content": {
                        "application/json" : {
                            "example" : {
                                "email" : "habil@gmail.com",
                                "username" : "habil",
                                "password" : "habil123",
                                "no_telepon" : "085238586875"
                            }
                        }
                    }
                },
                "responses": {
                    "201" : {
                        "description": "response if register succes",
                        "content": {
                           "application/json" : {
                                "example" : {
                                    "msg" : "register success",
                                    "data" : {
                                        "id" : "2345",
                                        "email" : "habil@gmail.com",
                                        "username" : "habil123",
                                        "no_telepon" : "085238586875",
                                        "acces_token" : "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c",
                                        "refresh_token" : "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
                                    }
                                }
                           }
                        }
                    },
                    "400" : {
                        "description": "response if email already exsist",
                        "content": {
                           "application/json" : {
                                "example" : {
                                    "msg" : "email already exist"
                                }
                           }
                        }
                    }
                }
            }
        },
        "/user/login" : {
            "post" : {
                "tags": ["user/auth"],
                "requestBody" : {
                    "content": {
                        "application/json" : {
                            "example" : {
                                "textBody" : "habil@gmail.com or 085238586875",
                                "password" : "habil123"
                            }
                        }
                    }
                },
                "responses": {
                    "200" : {
                        "description": "response if login succes,role on response can use for redirect page user like servant or vendee",
                        "content": {    
                           "application/json" : {
                                "example" : {                                  
                                    "msg" : "login success",
                                    "data" : {
                                             "role" : "servant",
                                             "acces_token" : "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c",
                                             "refresh_token" : "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
                                             }
                                }
                           }
                        }
                    },
                    "400" : {
                        "description": "response if email already exsist",
                        "content": {
                           "application/json" : {
                                "example" : {
                                    "msg" : "email or password wrong"
                                }
                           }
                        }
                    }
                }
            }
        },
...
        "/auth/select/role" : {
            "summary": "update or add role to user",
            "put" : {
                "parameters": [{
                    "$ref": "#/components/parameters/Authorization-Header"
                }],
                "tags": ["user/role"],
                "requestBody" : {
                    "content": {
                        "application/json" : {
                            "examples": {
                                "servantRole" : {
                                    "value" : {
                                        "role" : "servant"
                                    }
                                },
                                "vendeeRole" : {
                                    "value" : {
                                        "role" : "vendee"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200" : {
                        "description": "response if update role user succes",
                        "content": {
                            "application/json" : {
                                "example" : {
                                    "msg" : "update role success",
                                    "data" : {
                                        "id" : "2345",
                                        "username" : "habil",
                                        "email" : "habil@gmail.com",
                                        "role" : "servant or vendee"
                                    }
                                }
                            }
                        }
                    },
                    "400" : {
                        "description": "response if role user is exist,one user just can update role only one",
                        "content": {
                            "application/json" : {
                                "example" : {
                                    "msg" : "role user is exist"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "parameters": {
           "Authorization-Header" : {
                "in": "header",
                "name": "Authorization",
                "required": true,
                "description": "jwt acces token for Authorization user,if token is invalid response will be return Anauthorized(401)",
                "schema" : {
                    "type": "string",
                    "example" : "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
                }
            },
...
        }
    }
}